MAINT: Detach calibration from concentration analysis for better readibility.
